demo_frontend/requirements.txt
streamlit==1.41.0
sseclient==0.0.27
python-dotenv==1.0.1
openai==0.28